Adds planets via a loop, with misc textures.
This needs to be cleaned up before merging. Add more textures and provide a way to choose a random texture. Creating the sun should be placed outside of the viewDidLoadMethod.

Planets should be given their own class so creating them is handled outside of the view controller.

Contains a bug where planets will disappear on their trek on the far side of the sun.